requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
beautifulsoup4>=4.9.3
httpx>=0.24.0
markdownify>=0.11.6
python-dotenv>=0.19.0
googlesearch-python>=1.2.3
pytest>=7.0.0
fake-useragent>=1.1.3
jinja2>=3.0.0
